go testfixtures: move package to internal
