+++ internal/testfixtures/custom_type.go
