+++ livesql/example_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/samsarahq/thunder/livesql"
 	"github.com/samsarahq/thunder/reactive"
 	"github.com/samsarahq/thunder/sqlgen"
-	"github.com/samsarahq/thunder/testfixtures"
+	"github.com/samsarahq/thunder/internal/testfixtures"
 )
 
 type Cat struct {
